fixed meal plan updating

+++ IIS-Projekat/IIS-Projekat/Models/DTOs/Food/NewFoodDTO.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
         public string Category { get; set; }
-        public int Calories { get; set; }
+        public float Calories { get; set; }
         public List<int> AllergyIds { get; set; } = new List<int>();
         public NutrientTableDTO NutrientTable { get; set; }
     }
